add test package for new generators
