+++ recipes/qt/5.x.x/test_package/CMakeLists.txt
@@ -1,23 +1,9 @@
-cmake_minimum_required(VERSION 3.1.0)
-
-set(CMAKE_CXX_STANDARD 11)
-
-project(test_package)
-
-cmake_policy(SET CMP0020 NEW)
-
-include(${CMAKE_BINARY_DIR}/conanbuildinfo.cmake)
-conan_set_vs_runtime()
-conan_set_libcxx()
-conan_output_dirs_setup()
+cmake_minimum_required(VERSION 3.8)
+project(test_package LANGUAGES CXX)
 
 find_package(Qt5 COMPONENTS Core Network Sql Concurrent Xml REQUIRED CONFIG)
 
-set(CMAKE_AUTOMOC ON)
-set(CMAKE_AUTORCC ON)
-
-set(SOURCES test_package.cpp greeter.h example.qrc)
-
-add_executable(${PROJECT_NAME} WIN32 ${SOURCES})
-
-target_link_libraries(${PROJECT_NAME} Qt5::Core Qt5::Network Qt5::Sql Qt5::Concurrent Qt5::Xml)
+add_executable(${PROJECT_NAME} WIN32 test_package.cpp greeter.h example.qrc)
+target_link_libraries(${PROJECT_NAME} PRIVATE Qt5::Core Qt5::Network Qt5::Sql Qt5::Concurrent Qt5::Xml)
+target_compile_features(${PROJECT_NAME} PRIVATE cxx_std_11)
+set_target_properties(${PROJECT_NAME} PROPERTIES AUTOMOC ON AUTORCC ON)
