+++ recipes/qt/5.x.x/test_v1_package/test_package.pro
@@ -1,8 +1,8 @@
-SOURCES += test_package.cpp
+SOURCES += ../test_package/test_package.cpp
 
-HEADERS += greeter.h
+HEADERS += ../test_package/greeter.h
 
-RESOURCES = example.qrc
+RESOURCES = ../test_package/example.qrc
 
 QT -= gui
 QT += network sql concurrent xml
